grpuptableのidをint->bigintに変更

+++ pkg/db/create/create.go
@@ -7,7 +7,7 @@ import (
 
 var sql_stm []string = []string{
 	`create table IF NOT EXISTS "user" (user_id bigint PRIMARY KEY, nickname text NOT NULL, username text UNIQUE, password text)`,
-	`create table IF NOT EXISTS "group" (group_id int PRIMARY KEY, group_name text)`,
+	`create table IF NOT EXISTS "group" (group_id bigint PRIMARY KEY, group_name text)`,
 	`create table IF NOT EXISTS "user_group" (user_id int, group_id int, PRIMARY KEY(user_id, group_id), FOREIGN KEY(user_id) REFERENCES "user"(user_id), FOREIGN KEY(group_id) REFERENCES "group"(group_id))`,
 }
